+++ basculehttp/log_test.go
@@ -19,6 +19,7 @@ package basculehttp
 
 import (
 	"context"
+	"net/http"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -35,4 +36,37 @@ func TestGetZapLogger(t *testing.T) {
 	assert.NotPanics(t, func() {
 		result.Log("msg", "testing", "error", "nope", "level", "debug")
 	})
+}
+
+func TestSanitizeHeaders(t *testing.T) {
+	testCases := []struct {
+		Description string
+		Input       http.Header
+		Expected    http.Header
+	}{
+		{
+			Description: "Filtered",
+			Input:       http.Header{"Authorization": []string{"Basic xyz"}, "HeaderA": []string{"x"}},
+			Expected:    http.Header{"HeaderA": []string{"x"}, "Authorization-Type": []string{"Basic"}},
+		},
+		{
+			Description: "Handled human error",
+			Input:       http.Header{"Authorization": []string{"BasicXYZ"}, "HeaderB": []string{"y"}},
+			Expected:    http.Header{"HeaderB": []string{"y"}},
+		},
+		{
+			Description: "Not a perfect system",
+			Input:       http.Header{"Authorization": []string{"MySecret IWantToLeakIt"}},
+			Expected:    http.Header{"Authorization-Type": []string{"MySecret"}},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.Description, func(t *testing.T) {
+			assert := assert.New(t)
+			actual := sanitizeHeaders(tc.Input)
+			assert.Equal(tc.Expected, actual)
+		})
+
+	}
 }