+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Update setLogger Authorization header filtering logic. [#111](https://github.com/xmidt-org/bascule/pull/111)
 
 ## [v0.10.1]
 - Added raw parsers for bearer acquirer. [#110](https://github.com/xmidt-org/bascule/pull/110)